[abi] Handle missing Vec in no_std environment
We really need a further abstraction to handle memory types.
Right now they life in ink_core but we shouldn't depend on ink_core from within ink_abi.
So we want to extract the memory sub module out of ink_core into its own crate that we can from then on depend on.